--- sdk/core/System.ClientModel/src/Internal/Logging/LoggingEventIds.cs
@@ -1,24 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-namespace System.ClientModel.Internal;
-
-internal class LoggingEventIds
-{
-    public const int RequestEvent = 1;
-    public const int RequestContentEvent = 2;
-    public const int ResponseEvent = 5;
-    public const int ResponseContentEvent = 6;
-    public const int ResponseDelayEvent = 7;
-    public const int ErrorResponseEvent = 8;
-    public const int ErrorResponseContentEvent = 9;
-    public const int RequestRetryingEvent = 10;
-    public const int ResponseContentBlockEvent = 11;
-    public const int ErrorResponseContentBlockEvent = 12;
-    public const int ResponseContentTextEvent = 13;
-    public const int ErrorResponseContentTextEvent = 14;
-    public const int ResponseContentTextBlockEvent = 15;
-    public const int ErrorResponseContentTextBlockEvent = 16;
-    public const int RequestContentTextEvent = 17;
-    public const int ExceptionResponseEvent = 18;
-}
