+++ sdk/core/System.ClientModel/src/Internal/Logging/PipelineRetryLogger.cs
@@ -1,0 +1,17 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System.ClientModel.Primitives;
+using Microsoft.Extensions.Logging;
+
+namespace System.ClientModel.Internal;
+
+internal class PipelineRetryLogger
+{
+    private readonly ILogger<ClientRetryPolicy>? _logger;
+
+    public PipelineRetryLogger(ILoggerFactory? loggerFactory)
+    {
+        _logger = loggerFactory?.CreateLogger<ClientRetryPolicy>() ?? null;
+    }
+}
